Searcher Logic Extraction to SearchGo
